Paragraph on API dynamics to address issue #5
I didn't at a new section because there are a few hints on dynamic all over the doc already and especially here there was already text that talked about initiate and listen. So I thought that was a good fit here. However, if you want to rather have this in a separate section we have to probably go though the whole doc again and see what we already have said somewhere....